Ng2 Migration: Moving further code from angularOrcidOriginal to its own component
