+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/angularOrcidOriginal.js
@@ -18,360 +18,16 @@
 /*
  * Structure of this file:
  * 
- *  - 1 - Utility functions
- *  - 2 - Groupings logic
- *  - 3 - Angular Services
  *  - 4 - Angular Controllers
- *  - 5 - Angular Filters
  *  - 6 - Angular Directives
- *  - 7 - Angular Multiselect Module
  *  
  */
 
 /*******************************************************************************
- * CONTROLLERS
+ * 4 - Angular Controllers
 *******************************************************************************/
 
-angular.module('orcidApp').controller('DeactivateAccountCtrl', ['$scope', '$compile', function ($scope, $compile) {
-    $scope.sendDeactivateEmail = function() {
-        orcidGA.gaPush(['send', 'event', 'Disengagement', 'Deactivate_Initiate', 'Website']);
-        $.ajax({
-            url: getBaseUri() + '/account/send-deactivate-account.json',
-            dataType: 'text',
-            success: function(data) {
-                $scope.primaryEmail = data;
-                $.colorbox({
-                    html : $compile($('#deactivate-account-modal').html())($scope)
-                });
-                $scope.$apply();
-                $.colorbox.resize();
-            }
-        }).fail(function() {
-            // something bad is happening!
-            console.log("error with change DeactivateAccount");
-        });
-    };
 
-    $scope.closeModal = function() {
-        $.colorbox.close();
-    };
-}]);
-
-angular.module('orcidApp').controller('DeprecateAccountCtrl', ['$scope', '$compile', '$rootScope', 'emailSrvc', function ($scope, $compile, $rootScope, emailSrvc) {
-    $scope.emailSrvc = emailSrvc;
-    $scope.getDeprecateProfile = function() {
-        $.ajax({
-            url: getBaseUri() + '/account/deprecate-profile.json',
-            dataType: 'json',
-            success: function(data) {
-                $scope.deprecateProfilePojo = data;
-                $scope.$apply();
-            }
-        }).fail(function() {
-            console.log("An error occurred preparing deprecate profile");
-        });
-    };
-    
-    $scope.getDeprecateProfile();
-    
-    $scope.deprecateORCID = function() {
-        $.ajax({
-            url: getBaseUri() + '/account/validate-deprecate-profile.json',
-            dataType: 'json',
-            data: angular.toJson($scope.deprecateProfilePojo),
-            type: 'POST',
-            contentType: 'application/json;charset=UTF-8',
-            success: function(data) {
-                $scope.deprecateProfilePojo = data;
-                if (data.errors.length > 0) {
-                    $scope.$apply();
-                } else {
-                    $.colorbox({
-                        html : $compile($('#confirm-deprecate-account-modal').html())($scope),
-                        escKey:false,
-                        overlayClose:true,
-                        close: '',
-                        });
-                }
-                $scope.$apply();
-                $.colorbox.resize();
-            }
-        }).fail(function() {
-            // something bad is happening!
-            console.log("error with change DeactivateAccount");
-        });
-    };
-    
-    $scope.submitModal = function() {
-        $.ajax({
-            url: getBaseUri() + '/account/confirm-deprecate-profile.json',
-            type: 'POST',
-            data: angular.toJson($scope.deprecateProfilePojo),
-            contentType: 'application/json;charset=UTF-8',
-            dataType: 'json',
-            success: function(data) {                
-                emailSrvc.getEmails(function(emailData) {
-                    $rootScope.$broadcast('rebuildEmails', emailData);
-                });
-                $.colorbox({
-                    html : $compile($('#deprecate-account-confirmation-modal').html())($scope),
-                    escKey:false,
-                    overlayClose:true,
-                    close: '',
-                    onClosed: function(){ $scope.deprecateProfilePojo = null; $scope.$apply(); },
-                    });
-                $scope.$apply();
-                $.colorbox.resize();
-            }
-        }).fail(function() {
-            // something bad is happening!
-            console.log("error confirming account deprecation");
-        });
-    };
-
-    $scope.closeModal = function() {
-        $.colorbox.close();
-    };
-    
-}]);
-
-angular.module('orcidApp').controller('2FAStateCtrl', ['$scope', '$compile', function ($scope, $compile) {
-    
-    $scope.check2FAState = function() {
-        $.ajax({
-            url: getBaseUri() + '/2FA/status.json',
-            dataType: 'json',
-            success: function(data) {
-                $scope.update2FAStatus(data);
-            }
-        }).fail(function() {
-            console.log("An error occurred checking user's 2FA state");
-        });
-    };
-    
-    $scope.enable2FA = function() {
-        window.location.href = getBaseUri() + '/2FA/setup';
-    }
-    
-    $scope.disable2FA = function() {
-        $.ajax({
-            url: getBaseUri() + '/2FA/disable.json',
-            dataType: 'json',
-            type: 'POST',
-            success: function(data) {               
-                $scope.update2FAStatus(data);
-            }
-        }).fail(function() {
-            console.log("An error occurred disabling user's 2FA");
-        });
-    }
-    
-    $scope.update2FAStatus = function(status) {
-        $scope.showEnabled2FA = status.enabled;
-        $scope.showDisabled2FA = !status.enabled;
-        $scope.$apply();
-    }
-}]);
-
-angular.module('orcidApp').controller('2FASetupCtrl', ['$scope', '$compile', function ($scope, $compile) {
-    
-    $scope.startSetup = function() {
-        $.ajax({
-            url: getBaseUri() + '/2FA/QRCode.json',
-            dataType: 'json',
-            success: function(data) {
-                $("#2FA-QR-code").attr("src", data.url);
-                $scope.showSetup2FA = true;
-                $scope.showQRCode = true;
-                $scope.showTextCode = false;
-                $scope.show2FARecoveryCodes = false;
-                
-                $.ajax({
-                    url: getBaseUri() + '/2FA/register.json',
-                    dataType: 'json',
-                    success: function(data) {
-                        $scope.twoFactorAuthRegistration = data;
-                        $scope.$apply();
-                    }
-                }).fail(function(err) {
-                    console.log("An error occurred getting 2FA registration object");
-                });
-            }
-        }).fail(function(err) {
-            console.log("An error occurred getting user's 2FA QR code");
-        });
-        
-    };
-    
-    $scope.cancel2FASetup = function() {
-        window.location.href = getBaseUri() + "/account";
-    }
-    
-    $scope.done = function() {
-        window.location.href = getBaseUri() + "/account";
-    }
-    
-    $scope.sendVerificationCode = function() {
-        $.ajax({
-            url: getBaseUri() + '/2FA/register.json',
-            dataType: 'json',
-            data: angular.toJson($scope.twoFactorAuthRegistration),
-            contentType: 'application/json;charset=UTF-8',
-            type: 'POST',
-            success: function(data) {               
-                if (data.valid) {
-                    $scope.showSetup2FA = false;
-                    $scope.show2FARecoveryCodes = true;
-                    $scope.recoveryCodes = data.backupCodes;
-                    $scope.showInvalidCodeError=false;
-                } else {
-                    $scope.showInvalidCodeError=true;
-                }
-                $scope.$apply();
-            }
-        }).fail(function(xhr, status, error) {
-            var err = eval("(" + xhr.responseText + ")");
-            alert(err.Message);
-        });
-    }
-    
-    $('#getTextCode').click(function() {
-        $.ajax({
-            url: getBaseUri() + '/2FA/secret.json',
-            dataType: 'json',
-            success: function(data) {
-                $scope.textCodeFor2FA = data.secret;
-                $scope.showTextCode = true;
-                $scope.showQRCode = false;
-                $scope.$apply();
-            }
-        }).fail(function(err) {
-            console.log("An error occurred getting 2FA secret");
-        });
-    });
-    
-    $scope.copyRecoveryCodes = function() {
-        var recoveryCodesString = getRecoveryCodesString();
-        
-        if (window.clipboardData) { // for IE
-            window.clipboardData.setData("Text", recoveryCodesString);        
-        } else {
-            var temp = $('<div />');
-            temp.text(recoveryCodesString);
-            temp.css({
-                position: "absolute",
-                left:     "-1000px",
-                top:      "-1000px",
-            });
-            
-            $('body').append(temp);
-            
-            var range = document.createRange();
-            range.selectNodeContents(temp.get(0));
-            
-            var selection = window.getSelection();
-            selection.removeAllRanges();
-            selection.addRange(range);
-            
-            var copied = document.execCommand('copy', false, null);
-            alert(recoveryCodesString);
-            if (!copied) {
-                console.log("An error occurred copying recovery codes");
-            }
-            
-            temp.remove();
-        }
-    }   
-    
-    $scope.downloadRecoveryCodes = function() {
-        var recoveryCodesString = getRecoveryCodesString();
-        var element = document.createElement('a');
-        element.setAttribute('href', 'data:text/plain;charset=utf-8,' + encodeURIComponent(recoveryCodesString));
-        element.setAttribute('download', 'recovery-codes.txt');
-    
-        element.style.display = 'none';
-        document.body.appendChild(element);
-    
-        element.click();
-    
-        document.body.removeChild(element);
-    }
-    
-    function getRecoveryCodesString() {
-        var recoveryCodesString = "";
-        for (var i = 0; i < $scope.recoveryCodes.length; i++) {
-            recoveryCodesString += $scope.recoveryCodes[i] + "\n";
-        }
-        return recoveryCodesString;
-    }
-    
-}]);
-
-
-angular.module('orcidApp').controller('SecurityQuestionEditCtrl', ['$scope', '$compile', function ($scope, $compile) {
-    $scope.errors = null;
-    $scope.password = null;
-    $scope.securityQuestions = [];
-
-    $scope.getSecurityQuestion = function() {
-        $.ajax({
-            url: getBaseUri() + '/account/security-question.json',
-            dataType: 'json',
-            success: function(data) {               
-                $scope.securityQuestionPojo = data;
-                $scope.$apply();
-            }
-        }).fail(function() {
-            // something bad is happening!
-            console.log("error with security question.json");
-        });
-    };
-
-    $scope.getSecurityQuestion();
-
-    $scope.checkCredentials = function() {
-        $scope.password=null;
-        if(orcidVar.isPasswordConfirmationRequired){
-            $.colorbox({
-                html: $compile($('#check-password-modal').html())($scope)
-            });
-            $.colorbox.resize();
-        }
-        else{
-            $scope.submitModal();
-        }
-    };
-
-    $scope.submitModal = function() {
-        $scope.securityQuestionPojo.password=$scope.password;
-        $.ajax({
-            url: getBaseUri() + '/account/security-question.json',
-            type: 'POST',
-            data: angular.toJson($scope.securityQuestionPojo),
-            contentType: 'application/json;charset=UTF-8',
-            dataType: 'json',
-            success: function(data) {
-                
-                if(data.errors.length != 0) {
-                    $scope.errors=data.errors;
-                } else {
-                    $scope.errors=null;
-                }
-                $scope.getSecurityQuestion();
-                $scope.$apply();
-            }
-        }).fail(function() {
-            // something bad is happening!
-            console.log("error with security question");
-        });
-        $scope.password=null;
-        $.colorbox.close();
-    };
-
-    $scope.closeModal = function() {
-        $.colorbox.close();
-    };
-}]);
 
 angular.module('orcidApp').controller('RegistrationCtrl', ['$scope', '$compile', 'commonSrvc', 'vcRecaptchaService', function ($scope, $compile, commonSrvc, vcRecaptchaService) {
     $scope.privacyHelp = {};
